requirements.txt
# requirements.txt
google-generativeai==0.8.5
langchain-google-genai==2.0.10
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
faiss-cpu==1.11.0
python-dotenv==0.9.9
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-auth==2.40.2
numpy==1.26.4

# optional??
tqdm==4.67.1